Update docker publishing docs, remove PubSub instructions
Introduce a 'namespace' parameter so you don't need to change the pom to
deploy to your own namespace.

Removed the pubsub documentation since that code is not currently enabled,
and will likely be replaced with a different approach where the service
doesn't connect to PubSub 'manually' but through the proxy.